+++ architect_py/tests/test_sync_client.py
@@ -1,14 +1,12 @@
-import pytest
 from dotenv import load_dotenv
 
 from architect_py import Client
-from architect_py.tests.conftest import TestEnvironment
+from architect_py.tests.conftest import GetEnvironment
 
 
-@pytest.mark.asyncio
-async def test_sync_client():
+def test_sync_client():
     load_dotenv()
-    test_env = TestEnvironment.from_env()
+    test_env = GetEnvironment.from_env()
     client = Client(
         api_key=test_env.api_key,
         api_secret=test_env.api_secret,
@@ -21,3 +19,22 @@ async def test_sync_client():
 
     assert symbols is not None
     assert len(symbols) > 20
+
+    client.close()
+
+
+def test_creation():
+    load_dotenv()
+    test_env = GetEnvironment.from_env()
+    client = Client(
+        api_key=test_env.api_key,
+        api_secret=test_env.api_secret,
+        paper_trading=True,
+        endpoint="app.architect.co",
+    )
+
+    symbols = client.list_symbols(marketdata="CME")
+
+    assert symbols is not None
+    assert len(symbols) > 20
+    client.close()