+++ examples/stream_l2_marketdata.py
@@ -33,5 +33,7 @@ async def main():
     venue = "CME"
     await print_l2_book(c, market_symbol, venue=venue)
 
+    await c.close()
+
 
 asyncio.run(main())